abstracted footer, removed a lot of inline styling, restructured how blog works
