+++ src/components/Footer.js
@@ -1,0 +1,15 @@
+import React from 'react';
+
+const Footer = () => (
+  <footer>
+    <a rel='license' href='http://creativecommons.org/licenses/by-sa/4.0/'>
+      <img
+        style={{ borderWidth: 0 }}
+        src='https://i.creativecommons.org/l/by-sa/4.0/80x15.png'
+        alt='Creative Commons License'
+      />
+    </a>
+  </footer>
+);
+
+export default Footer;
