+++ src/pages/404.js
@@ -1,6 +1,5 @@
 import React from 'react';
-
-import Layout from '../components/layout';
+import Layout from '../components/Layout';
 import SEO from '../components/SEO';
 
 const NotFoundPage = () => (
