--- src/pages/blog/indexBAK.js
@@ -1,48 +1,0 @@
-import React from 'react';
-import { graphql, Link } from 'gatsby';
-import Layout from '../../components/layout';
-
-const BlogIndex = ({ data }) => {
-  const { edges: posts } = data.allMdx;
-
-  return (
-    <Layout>
-      {posts.map(({ node }) => {
-        const { title, author } = node.frontmatter;
-        return (
-          <div key={node.id}>
-            <header>
-              <div>{title}</div>
-              <div>Posting By {author}</div>
-            </header>
-            <p>{node.excerpt}</p>
-            <Link to={node.fields.slug}>View Article</Link>
-            <hr />
-          </div>
-        );
-      })}
-    </Layout>
-  );
-};
-
-export default BlogIndex;
-
-export const pageQuery = graphql`
-  query blogIndex {
-    allMdx {
-      edges {
-        node {
-          id
-          excerpt
-          fields {
-            slug
-          }
-          frontmatter {
-            title
-            author
-          }
-        }
-      }
-    }
-  }
-`;
