+++ src/templates/BrowsePosts.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import { Link } from 'gatsby/index';
-import Layout from '../components/layout';
+import { Link } from 'gatsby';
+import Layout from '../components/Layout';
 
 const BrowsePosts = ({ pageContext }) => {
   const { groupedPosts, group, page } = pageContext;
